+++ ktrain/text/predictor.py
@@ -1,6 +1,6 @@
 from ..imports import *
 from ..predictor import Predictor
-from .preprocessor import TextPreprocessor
+from .preprocessor import TextPreprocessor, TransformersPreprocessor
 from .. import utils as U
 
 class TextPredictor(Predictor):
@@ -46,6 +46,9 @@ class TextPredictor(Predictor):
         #    treat_multilabel = True
         texts = self.preproc.preprocess(texts)
         preds = self.model.predict(texts)
+        if U.is_huggingface(model=self.model):
+            # convert logits to probabilities for Hugging Face models
+            preds = activations.softmax(tf.convert_to_tensor(preds)).numpy()
         result =  preds if return_proba else [self.c[np.argmax(pred)] for pred in preds] 
         if multilabel:
             result =  [list(zip(self.c, r)) for r in result]
@@ -63,12 +66,14 @@ class TextPredictor(Predictor):
         return self.predict(texts, return_proba=True)
 
 
-    def explain(self, doc, truncate_len=512):
+    def explain(self, doc, truncate_len=512, all_targets=False):
         """
         Highlights text to explain prediction
         Args:
             doc (str): text of documnet
             truncate_len(int): truncate document to this many words
+            all_targets(bool):  If True, show visualization for
+                                each target.
         """
         try:
             import eli5
@@ -79,6 +84,7 @@ class TextPredictor(Predictor):
             warnings.warn(msg)
             return
 
+        prediction = [self.predict(doc)] if not all_targets else None
 
         if not isinstance(doc, str): raise Exception('text must of type str')
         if self.preproc.is_nospace_lang():
@@ -87,7 +93,7 @@ class TextPredictor(Predictor):
         doc = ' '.join(doc.split()[:truncate_len])
         te = TextExplainer(random_state=42)
         _ = te.fit(doc, self.predict_proba)
-        return te.show_prediction(target_names=self.preproc.get_classes())
+        return te.show_prediction(target_names=self.preproc.get_classes(), targets=prediction)
 
 
     def analyze_valid(self, val_tup, print_report=True, multilabel=None):
@@ -118,3 +124,21 @@ class TextPredictor(Predictor):
             cm_func = confusion_matrix
         cm =  confusion_matrix(y_true,  y_pred)
         return cm
+
+
+    def save(self, fpath):
+
+        if isinstance(self.preproc, TransformersPreprocessor):
+            if os.path.isfile(fpath):
+                raise ValueError(f'There is an existing file named {fpath}. ' +\
+                                  'Please use dfferent value for fpath.')
+            elif not os.path.exists(fpath):
+                os.mkdir(fpath)
+            self.model.save_pretrained(fpath)
+            fname_preproc = fpath+'.preproc'
+            with open(fname_preproc, 'wb') as f:
+                pickle.dump(self.preproc, f)
+        else:
+            super().save(fpath)
+        return
+